test(host): split the #342 matrix by contract area (#342)
One 1,392-line file held sixteen numbered scenarios spanning five
unrelated contracts, so a failure told you "the matrix broke" and any
change touched a file everyone else was also touching.

Now five files, each falsifying one contract and saying so in its own
docstring:

  test_batch_public_api.py       402  submission surface + map expansion
  test_batch_pause_lifecycle.py  247  one child: pause, answer, resume, loop
  test_batch_views_and_watch.py  348  what a detached operator can SEE
  test_batch_stop_and_races.py   338  stop, closed admission, who wins
  test_batch_sync_parity.py      136  the synchronous mirrors

Shared setup stays small and explicit. `conftest.py` holds ONLY the two
things pytest must resolve by name (`home`, `ledger`) — a fixture cannot
be imported without making every test that uses it look like a
redefinition. The callable helpers live in `_batch_interrupt.py` and are
imported by name, so each file's header still lists exactly what it uses.
Every pre-existing ticket suite defines its own module-level `home`, which
takes precedence, so none of them changes behavior.

Regrouped rather than moved verbatim: the two tolerance tests left the
"mixed batch" class for a `TestClosedAdmission` class in the stop/races
file, and the sync store-mirror test left the release-window class for the
sync-parity file.

No test was added, removed, or weakened: 83 tests before, 83 after.
`uv run pytest -m host_batch_interrupt -q` 83 passed.
